+++ apps/api/src/decorators/non-empty-trimmed-string.decorator.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty } from 'class-validator'
+import { TrimmedString } from '@/decorators/trimmed-string.decorator'
+import { applyDecorators } from '@nestjs/common'
+
+export function NonEmptyTrimmedString() {
+  return applyDecorators(TrimmedString(), IsNotEmpty())
+}