+++ apps/api/src/secret/dto/create.secret/create.secret.ts
@@ -1,22 +1,18 @@
 import 'reflect-metadata'
-import { Transform, Type } from 'class-transformer'
+import { Type } from 'class-transformer'
 import {
   IsArray,
-  IsNotEmpty,
   IsOptional,
   IsString,
   Length,
   ValidateNested
 } from 'class-validator'
-import { TrimString } from '@/decorators/trim-string.decorator'
+import { NonEmptyTrimmedString } from '@/decorators/non-empty-trimmed-string.decorator'
 
 export class CreateSecret {
-  @IsString()
-  @IsNotEmpty()
-  @TrimString()
+  @NonEmptyTrimmedString()
   name: string
 
-  @IsString()
   @IsOptional()
   @Length(0, 100)
   note?: string
@@ -33,13 +29,9 @@ export class CreateSecret {
 }
 
 class Entry {
-  @IsString()
-  @IsNotEmpty()
-  @TrimString()
+  @NonEmptyTrimmedString()
   environmentSlug: string
 
-  @IsString()
-  @IsNotEmpty()
-  @TrimString()
+  @NonEmptyTrimmedString()
   value: string
 }