+++ apps/platform/src/components/dashboard/variable/editVariableSheet/index.tsx
@@ -2,7 +2,6 @@
 
 import { useCallback, useState } from 'react'
 import { toast } from 'sonner'
-import type { UpdateVariableRequest } from '@keyshade/schema'
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
@@ -23,6 +22,7 @@ import {
   selectedVariableAtom,
   variablesOfProjectAtom
 } from '@/store'
+import { useHttp } from '@/hooks/use-http'
 
 export default function EditVariablSheet() {
   const [isEditVariableOpen, setIsEditVariableOpen] =
@@ -37,6 +37,20 @@ export default function EditVariablSheet() {
     name: selectedVariableData?.variable.name,
     note: selectedVariableData?.variable.note || ''
   })
+  const [isLoading, setIsLoading] = useState<boolean>(false)
+
+  const updateVariable = useHttp(() =>
+    ControllerInstance.getInstance().variableController.updateVariable({
+      variableSlug: selectedVariableData!.variable.slug,
+      name:
+        requestData.name === selectedVariableData!.variable.name ||
+        requestData.name === ''
+          ? undefined
+          : requestData.name,
+      note: requestData.note === '' ? undefined : requestData.note,
+      entries: undefined
+    })
+  )
 
   const handleClose = useCallback(() => {
     setIsEditVariableOpen(false)
@@ -49,79 +63,57 @@ export default function EditVariablSheet() {
     }))
   }, [])
 
-  const updateVariable = useCallback(async () => {
-    if (!selectedVariableData) {
-      toast.error('No variable selected', {
-        description: (
-          <p className="text-xs text-red-300">
-            No variable selected. Please select a variable.
-          </p>
-        )
-      })
-      return
-    }
+  const handleUpdateVariable = useCallback(async () => {
+    if (selectedVariableData) {
+      setIsLoading(true)
+      toast.loading('Updating variable...')
 
-    const { variable } = selectedVariableData
+      try {
+        const { success, data } = await updateVariable()
 
-    const request: UpdateVariableRequest = {
-      variableSlug: variable.slug,
-      name:
-        requestData.name === variable.name || requestData.name === ''
-          ? undefined
-          : requestData.name,
-      note: requestData.note === '' ? undefined : requestData.note,
-      entries: undefined
-    }
+        if (success && data) {
+          toast.success('Variable edited successfully', {
+            description: (
+              <p className="text-xs text-emerald-300">
+                You successfully edited the variable
+              </p>
+            )
+          })
 
-    const { success, error, data } =
-      await ControllerInstance.getInstance().variableController.updateVariable(
-        request,
-        {}
-      )
-
-    if (success && data) {
-      toast.success('Variable edited successfully', {
-        description: (
-          <p className="text-xs text-emerald-300">
-            You successfully edited the variable
-          </p>
-        )
-      })
-
-      // Update the variable in the store
-      setVariables((prev) => {
-        const newVariables = prev.map((v) => {
-          if (v.variable.slug === variable.slug) {
-            return {
-              ...v,
-              variable: {
-                ...v.variable,
-                name: requestData.name || v.variable.name,
-                note: requestData.note || v.variable.note,
-                slug: data.variable.slug
+          // Update the variable in the store
+          setVariables((prev) => {
+            const newVariables = prev.map((v) => {
+              if (v.variable.slug === selectedVariableData.variable.slug) {
+                return {
+                  ...v,
+                  variable: {
+                    ...v.variable,
+                    name: requestData.name || v.variable.name,
+                    note: requestData.note || v.variable.note,
+                    slug: data.variable.slug
+                  }
+                }
               }
-            }
-          }
-          return v
-        })
-        return newVariables
-      })
-    }
-    if (error) {
-      toast.error('Something went wrong!', {
-        description: (
-          <p className="text-xs text-red-300">
-            Something went wrong while updating the variable. Check console for
-            more info.
-          </p>
-        )
-      })
-      // eslint-disable-next-line no-console -- we need to log the error
-      console.error('Error while updating variable: ', error)
-    }
+              return v
+            })
+            return newVariables
+          })
 
-    handleClose()
-  }, [selectedVariableData, requestData, handleClose, setVariables])
+          handleClose()
+        }
+      } finally {
+        setIsLoading(false)
+        toast.dismiss()
+      }
+    }
+  }, [
+    selectedVariableData,
+    updateVariable,
+    setVariables,
+    handleClose,
+    requestData.name,
+    requestData.note
+  ])
 
   return (
     <Sheet
@@ -145,12 +137,7 @@ export default function EditVariablSheet() {
             <Input
               className="col-span-3 h-[2.75rem]"
               id="name"
-              onChange={(e) => {
-                setRequestData((prev) => ({
-                  ...prev,
-                  name: e.target.value
-                }))
-              }}
+              onChange={(e) => updateRequestData(e.target.id, e.target.value)}
               placeholder="Enter the name of the variable"
               value={requestData.name}
             />
@@ -163,12 +150,7 @@ export default function EditVariablSheet() {
             <Textarea
               className="col-span-3 h-[2.75rem]"
               id="name"
-              onChange={(e) => {
-                setRequestData((prev) => ({
-                  ...prev,
-                  note: e.target.value
-                }))
-              }}
+              onChange={(e) => updateRequestData(e.target.id, e.target.value)}
               placeholder="Enter the note of the variable"
               value={requestData.note}
             />
@@ -178,7 +160,8 @@ export default function EditVariablSheet() {
           <SheetClose asChild>
             <Button
               className="font-semibold"
-              onClick={updateVariable}
+              disabled={isLoading}
+              onClick={handleUpdateVariable}
               variant="secondary"
             >
               Edit Variable