+++ apps/platform/src/components/shared/navbar/index.tsx
@@ -4,10 +4,9 @@ import { Search } from 'lucide-react'
 import { useCallback, useEffect, useState } from 'react'
 import { usePathname } from 'next/navigation'
 import Link from 'next/link'
-import { DropdownSVG } from '@public/svg/shared'
-import { SecretSVG, VariableSVG, EnvironmentSVG } from '@public/svg/dashboard'
-import type { User } from '@keyshade/schema'
 import { useAtomValue } from 'jotai'
+import { SecretSVG, VariableSVG, EnvironmentSVG } from '@public/svg/dashboard'
+import { DropdownSVG } from '@public/svg/shared'
 import SearchModel from './searchModel'
 import {
   DropdownMenu,
@@ -18,9 +17,8 @@ import {
   DropdownMenuTrigger
 } from '@/components/ui/dropdown-menu'
 import LineTab from '@/components/ui/line-tab'
-import { selectedProjectAtom } from '@/store'
-import { userAtom } from '@/store'
 import AvatarComponent from '@/components/common/avatar'
+import { selectedProjectAtom, userAtom } from '@/store'
 
 function Navbar(): React.JSX.Element {
   const [isOpen, setIsOpen] = useState<boolean>(false)